Extract admin-only CatalogPromotionType
