+++ src/Sylius/Bundle/AdminBundle/Form/Type/CatalogPromotionScopeType.php
@@ -14,55 +14,59 @@ declare(strict_types=1);
 namespace Sylius\Bundle\AdminBundle\Form\Type;
 
 use Sylius\Bundle\PromotionBundle\Form\Type\CatalogPromotionScopeType as BaseCatalogPromotionScopeType;
+use Sylius\Component\Promotion\Model\CatalogPromotionScopeInterface;
 use Symfony\Component\Form\AbstractType;
-use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
+use Symfony\Component\Form\Extension\Core\Type\HiddenType;
 use Symfony\Component\Form\FormBuilderInterface;
-use Twig\Environment;
+use Symfony\Component\Form\FormEvent;
+use Symfony\Component\Form\FormEvents;
 
 final class CatalogPromotionScopeType extends AbstractType
 {
-    /** @var array<string, string> */
-    private array $scopeTypes = [];
-
     /** @var array<string, string> */
     private array $scopeConfigurationTypes;
 
     /**
      * @param iterable<string, object> $scopeConfigurationTypes
      */
-    public function __construct(iterable $scopeConfigurationTypes, private Environment $twig)
+    public function __construct(iterable $scopeConfigurationTypes)
     {
         foreach ($scopeConfigurationTypes as $type => $formType) {
             $this->scopeConfigurationTypes[$type] = $formType::class;
-            $this->scopeTypes['sylius.form.catalog_promotion.scope.' . $type] = $type;
         }
     }
 
     public function buildForm(FormBuilderInterface $builder, array $options): void
     {
-        $builder
-            ->add('type', ChoiceType::class, [
-                'label' => 'sylius.ui.type',
-                'choices' => $this->scopeTypes,
-                'choice_attr' => function (?string $type) use ($builder): array {
-                    return [
-                        'data-configuration' => $this->twig->render(
-                            '@SyliusAdmin/CatalogPromotion/Scope/' . $type . '.html.twig',
-                            ['field' => $builder->create('configuration', $this->scopeConfigurationTypes[$type])->getForm()->createView()],
-                        ),
-                    ];
-                },
-            ])
-        ;
-    }
+        $builder->add('type', HiddenType::class);
 
-    public function getBlockPrefix(): string
-    {
-        return 'sylius_admin_catalog_promotion_scope';
+        $builder
+            ->addEventListener(FormEvents::PRE_SET_DATA, function (FormEvent $event): void {
+                $this->addScopeToForm($event);
+            })
+            ->addEventListener(FormEvents::PRE_SUBMIT, function (FormEvent $event): void {
+                $this->addScopeToForm($event);
+            })
+        ;
     }
 
     public function getParent(): string
     {
         return BaseCatalogPromotionScopeType::class;
+    }
+
+    private function addScopeToForm(FormEvent $event): void
+    {
+        $data = $event->getData();
+
+        if ($data === null) {
+            return;
+        }
+        $form = $event->getForm();
+
+        $dataType = $data instanceof CatalogPromotionScopeInterface ? $data->getType() : $data['type'];
+
+        $scopeConfigurationType = $this->scopeConfigurationTypes[$dataType];
+        $form->add('configuration', $scopeConfigurationType);
     }
 }
