+++ src/Sylius/Bundle/AdminBundle/Form/Type/CatalogPromotionType.php
@@ -15,34 +15,54 @@ namespace Sylius\Bundle\AdminBundle\Form\Type;
 
 use Sylius\Bundle\PromotionBundle\Form\Type\CatalogPromotionType as BaseCatalogPromotionType;
 use Symfony\Component\Form\AbstractType;
-use Symfony\Component\Form\Extension\Core\Type\CollectionType;
 use Symfony\Component\Form\FormBuilderInterface;
+use Symfony\UX\LiveComponent\Form\Type\LiveCollectionType;
 
 final class CatalogPromotionType extends AbstractType
 {
-    public function buildForm(FormBuilderInterface $builder, array $options): void
-    {
-        $builder->add('scopes', CollectionType::class, [
-            'label' => 'sylius.ui.scopes',
-            'entry_type' => CatalogPromotionScopeType::class,
-            'allow_add' => true,
-            'allow_delete' => true,
-            'by_reference' => false,
-            'required' => false,
-        ])
-        ->add('actions', CollectionType::class, [
-            'label' => 'sylius.ui.actions',
-            'entry_type' => CatalogPromotionActionType::class,
-            'allow_add' => true,
-            'allow_delete' => true,
-            'by_reference' => false,
-            'required' => false,
-        ]);
+    /**
+     * @param array<string, string> $scopeTypes
+     * @param array<string, string> $actionTypes
+     */
+    public function __construct(
+        private readonly array $scopeTypes,
+        private readonly array $actionTypes,
+    ) {
     }
 
-    public function getBlockPrefix(): string
+    /** @param array<string, mixed> $options */
+    public function buildForm(FormBuilderInterface $builder, array $options): void
     {
-        return 'sylius_admin_catalog_promotion';
+        $builder
+            ->add('scopes', LiveCollectionType::class, [
+                'entry_type' => CatalogPromotionScopeType::class,
+                'allow_add' => true,
+                'allow_delete' => true,
+                'by_reference' => false,
+                'button_add_type' => AddButtonType::class,
+                'button_add_options' => [
+                    'label' => 'sylius.ui.add_scope',
+                    'types' => $this->scopeTypes,
+                ],
+                'button_delete_options' => [
+                    'label' => false,
+                ],
+            ])
+            ->add('actions', LiveCollectionType::class, [
+                'entry_type' => CatalogPromotionActionType::class,
+                'allow_add' => true,
+                'allow_delete' => true,
+                'by_reference' => false,
+                'button_add_type' => AddButtonType::class,
+                'button_add_options' => [
+                    'label' => 'sylius.ui.add_action',
+                    'types' => $this->actionTypes,
+                ],
+                'button_delete_options' => [
+                    'label' => false,
+                ],
+            ])
+        ;
     }
 
     public function getParent(): string
