+++ src/Sylius/Bundle/AdminBundle/Resources/config/services/twig/component.xml
@@ -260,6 +260,16 @@
             />
         </service>
 
+        <service id="Sylius\Bundle\AdminBundle\Twig\Component\CatalogPromotion\FormComponent">
+            <argument type="service" id="form.factory" />
+            <argument>Sylius\Bundle\AdminBundle\Form\Type\CatalogPromotionType</argument>
+        </service>
+
+        <service id="Sylius\Bundle\AdminBundle\Twig\Component\Product\GenerateProductVariantsFormComponent">
+            <argument type="service" id="form.factory" />
+            <argument>Sylius\Bundle\ProductBundle\Form\Type\ProductGenerateVariantsType</argument>
+        </service>
+
         <service
             id="sylius_admin.twig.component.shipment.ship_form"
             class="Sylius\Bundle\AdminBundle\Twig\Component\Shipment\ShipFormComponent"
@@ -301,6 +311,5 @@
                 template="@SyliusAdmin/taxon/sections/tree.html.twig"
             />
         </service>
-
     </services>
 </container>
